pyproject.toml
[project]
name = "mqttactions"
version = "0.4.0"
description = "Toolkit to interact with MQTT to build simple action rules"
readme = "README.md"
license = { file = "LICENSE.txt" }
authors = [
    { name = "Manuel Stocker", email = "mensi@mensi.ch" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "fastapi>=0.111.0",
    "hatchling>=1.27.0",
    "paho-mqtt>=2.1.0",
    "suntime>=1.3.2",
    "timezonefinder>=6.5.9",
    "tzdata>=2025.2 ; sys_platform == 'win32'",
    "uvicorn>=0.30.0",
    "jinja2",
    "websockets",
    "pydantic",
]

[project.urls]
"Homepage" = "https://github.com/mensi/mqttactions"

[project.scripts]
mqttactions = "mqttactions.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
